+++ composeApp/src/commonMain/kotlin/ai/mcpdirect/studio/app/model/OpenAPIServerConfig.kt
@@ -4,6 +4,8 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 class OpenAPIServerConfig {
+    val id: Long = 0
+    val name: String = ""
     var status: Int = 1
 
     var url: String? = null