+++ composeApp/src/commonMain/kotlin/ai/mcpdirect/studio/app/model/repository/ToolRepository.kt
@@ -2,22 +2,17 @@ package ai.mcpdirect.studio.app.model.repository
 
 import ai.mcpdirect.mcpdirectstudioapp.currentMilliseconds
 import ai.mcpdirect.mcpdirectstudioapp.getPlatform
-import ai.mcpdirect.studio.app.UIState
 import ai.mcpdirect.studio.app.auth.authViewModel
 import ai.mcpdirect.studio.app.generalViewModel
-import ai.mcpdirect.studio.app.model.MCPServer
-import ai.mcpdirect.studio.app.model.aitool.*
-import ai.mcpdirect.studio.app.model.repository.StudioRepository.modifyMCPServerNameForStudio
-import ai.mcpdirect.studio.app.model.repository.StudioRepository.toolAgent
-import androidx.lifecycle.viewModelScope
+import ai.mcpdirect.studio.app.model.aitool.AIPortMCPServerConfig
+import ai.mcpdirect.studio.app.model.aitool.AIPortTool
+import ai.mcpdirect.studio.app.model.aitool.AIPortToolMaker
+import ai.mcpdirect.studio.app.model.aitool.AIPortToolMakerTemplate
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.update
-import kotlinx.coroutines.launch
 import kotlinx.coroutines.sync.Mutex
 import kotlinx.coroutines.sync.withLock
-import kotlin.collections.get
-import kotlin.collections.set
 import kotlin.time.Duration.Companion.seconds
 import kotlin.time.TimeMark
 import kotlin.time.TimeSource